:sparkles: request reviews from BoekenLiefde API
from now one we are using a second API to enhance the user experience. user tests concluded that they would like to have more information on a book (i.e. ratings and user reviews). luckily we've found a Dutch API that serves those. the API is from the BoekenLiefde website.