Feature/reviews and styling (#144)
* :sparkles: request reviews from BoekenLiefde API

from now one we are using a second API to enhance the user experience. user tests concluded that they would like to have more information on a book (i.e. ratings and user reviews). luckily we've found a Dutch API that serves those. the API is from the BoekenLiefde website.

* :sparkles: add more endpoints

* :art: add details tag

* :lipstick: svg arrow bullets > block

remove svg arrows functioning as bullet points and replace them with a utf-8 block

* :lipstick: fix some issues with the bullet points

* :art: reorder items

* :art: move availability to details element

* :art: add classes to each item for styling purposes

* :lipstick: add styling